refactor: MajorExceptionHandler -> MandatoryMajorSpecialCaseHandler 변경
